+++ flask-backend/app/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask, render_template, request
+
+app = Flask(__name__)
+app.config['DEBUG'] = True
+
+from app import routes
