+++ flask-backend/app/routes.py
@@ -1,0 +1,6 @@
+import flask
+from app import app
+
+@app.route("/")
+def my_index():
+    return flask.render_template("index.html", token="Flask Token")
