+++ flask-backend/app/static/react/asset-manifest.json
@@ -1,0 +1,23 @@
+{
+  "files": {
+    "main.css": "/static/react/css/main.5f361e03.chunk.css",
+    "main.js": "/static/react/js/main.c838d6cd.chunk.js",
+    "main.js.map": "/static/react/js/main.c838d6cd.chunk.js.map",
+    "runtime-main.js": "/static/react/js/runtime-main.8fa3ef7f.js",
+    "runtime-main.js.map": "/static/react/js/runtime-main.8fa3ef7f.js.map",
+    "js/2.cf1e2948.chunk.js": "/static/react/js/2.cf1e2948.chunk.js",
+    "js/2.cf1e2948.chunk.js.map": "/static/react/js/2.cf1e2948.chunk.js.map",
+    "../../templates/index.html": "/static/react/../../templates/index.html",
+    "css/main.5f361e03.chunk.css.map": "/static/react/css/main.5f361e03.chunk.css.map",
+    "js/2.cf1e2948.chunk.js.LICENSE.txt": "/static/react/js/2.cf1e2948.chunk.js.LICENSE.txt",
+    "media/logo.svg": "/static/react/media/logo.5d5d9eef.svg",
+    "precache-manifest.3064b550b2e09211425e4a4a2c8d3b76.js": "/static/react/precache-manifest.3064b550b2e09211425e4a4a2c8d3b76.js",
+    "service-worker.js": "/static/react/service-worker.js"
+  },
+  "entrypoints": [
+    "js/runtime-main.8fa3ef7f.js",
+    "js/2.cf1e2948.chunk.js",
+    "css/main.5f361e03.chunk.css",
+    "js/main.c838d6cd.chunk.js"
+  ]
+}
