+++ flask-backend/app/static/react/favicon.ico
