+++ flask-backend/app/static/react/manifest.json
