+++ flask-backend/app/static/react/precache-manifest.3064b550b2e09211425e4a4a2c8d3b76.js
@@ -1,0 +1,30 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "2a52fedcc96822797a53abbea819428d",
+    "url": "/static/react/../../templates/index.html"
+  },
+  {
+    "revision": "676fbc12ab166848f02a",
+    "url": "/static/react/css/main.5f361e03.chunk.css"
+  },
+  {
+    "revision": "336b1d4b5bca64a64fdf",
+    "url": "/static/react/js/2.cf1e2948.chunk.js"
+  },
+  {
+    "revision": "e88a3e95b5364d46e95b35ae8c0dc27d",
+    "url": "/static/react/js/2.cf1e2948.chunk.js.LICENSE.txt"
+  },
+  {
+    "revision": "676fbc12ab166848f02a",
+    "url": "/static/react/js/main.c838d6cd.chunk.js"
+  },
+  {
+    "revision": "642bffd405cc52de04d1",
+    "url": "/static/react/js/runtime-main.8fa3ef7f.js"
+  },
+  {
+    "revision": "5d5d9eefa31e5e13a6610d9fa7a283bb",
+    "url": "/static/react/media/logo.5d5d9eef.svg"
+  }
+]);
