+++ flask-backend/app/static/react/service-worker.js
@@ -14,7 +14,7 @@
 importScripts("https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js");
 
 importScripts(
-  "/static/react/precache-manifest.ef6dcd28f1e9e7e985db6e3e26ea3052.js"
+  "/static/react/precache-manifest.3064b550b2e09211425e4a4a2c8d3b76.js"
 );
 
 self.addEventListener('message', (event) => {
