+++ flask-backend/app/templates/index.html
@@ -1,0 +1,79 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8"/>
+    <link rel="icon" href="/static/react/favicon.ico"/>
+    <meta name="viewport" content="width=device-width,initial-scale=1"/>
+    <meta name="theme-color" content="#000000"/>
+    <meta name="description" content="Web site created using create-react-app"/>
+    <link rel="apple-touch-icon" href="/static/react/logo192.png"/>
+    <link rel="manifest" href="/static/react/manifest.json"/>
+    <title>React App</title>
+    <script>window.token = "{{token}}"</script>
+    <link href="/static/react/css/main.5f361e03.chunk.css" rel="stylesheet">
+</head>
+<body>
+<noscript>You need to enable JavaScript to run this app.</noscript>
+<div id="root"></div>
+<script>!function (e) {
+    function t(t) {
+        for (var n, a, f = t[0], l = t[1], c = t[2], p = 0, s = []; p < f.length; p++) a = f[p], Object.prototype.hasOwnProperty.call(o, a) && o[a] && s.push(o[a][0]), o[a] = 0;
+        for (n in l) Object.prototype.hasOwnProperty.call(l, n) && (e[n] = l[n]);
+        for (i && i(t); s.length;) s.shift()();
+        return u.push.apply(u, c || []), r()
+    }
+
+    function r() {
+        for (var e, t = 0; t < u.length; t++) {
+            for (var r = u[t], n = !0, f = 1; f < r.length; f++) {
+                var l = r[f];
+                0 !== o[l] && (n = !1)
+            }
+            n && (u.splice(t--, 1), e = a(a.s = r[0]))
+        }
+        return e
+    }
+
+    var n = {}, o = {1: 0}, u = [];
+
+    function a(t) {
+        if (n[t]) return n[t].exports;
+        var r = n[t] = {i: t, l: !1, exports: {}};
+        return e[t].call(r.exports, r, r.exports, a), r.l = !0, r.exports
+    }
+
+    a.m = e, a.c = n, a.d = function (e, t, r) {
+        a.o(e, t) || Object.defineProperty(e, t, {enumerable: !0, get: r})
+    }, a.r = function (e) {
+        "undefined" != typeof Symbol && Symbol.toStringTag && Object.defineProperty(e, Symbol.toStringTag, {value: "Module"}), Object.defineProperty(e, "__esModule", {value: !0})
+    }, a.t = function (e, t) {
+        if (1 & t && (e = a(e)), 8 & t) return e;
+        if (4 & t && "object" == typeof e && e && e.__esModule) return e;
+        var r = Object.create(null);
+        if (a.r(r), Object.defineProperty(r, "default", {
+            enumerable: !0,
+            value: e
+        }), 2 & t && "string" != typeof e) for (var n in e) a.d(r, n, function (t) {
+            return e[t]
+        }.bind(null, n));
+        return r
+    }, a.n = function (e) {
+        var t = e && e.__esModule ? function () {
+            return e.default
+        } : function () {
+            return e
+        };
+        return a.d(t, "a", t), t
+    }, a.o = function (e, t) {
+        return Object.prototype.hasOwnProperty.call(e, t)
+    }, a.p = "/static/react/";
+    var f = this["webpackJsonpreact-frontend"] = this["webpackJsonpreact-frontend"] || [], l = f.push.bind(f);
+    f.push = t, f = f.slice();
+    for (var c = 0; c < f.length; c++) t(f[c]);
+    var i = l;
+    r()
+}([])</script>
+<script src="/static/react/js/2.cf1e2948.chunk.js"></script>
+<script src="/static/react/js/main.c838d6cd.chunk.js"></script>
+</body>
+</html>
