+++ flask-backend/main.py
@@ -1,9 +1,4 @@
-import flask
+from app import app
 
-app = flask.Flask("__main__")
-
-@app.route("/")
-def my_index():
-    return flask.render_template("index.html", token="Flask Token")
-
-app.run(debug=True)
+if __name__ == '__main__':
+    app.run(debug =True)
