--- flask-backend/static/react/asset-manifest.json
@@ -1,22 +1,0 @@
-{
-  "files": {
-    "main.css": "/static/react/css/main.5f361e03.chunk.css",
-    "main.js": "/static/react/js/main.29a4e32b.chunk.js",
-    "main.js.map": "/static/react/js/main.29a4e32b.chunk.js.map",
-    "runtime-main.js": "/static/react/js/runtime-main.8fa3ef7f.js",
-    "runtime-main.js.map": "/static/react/js/runtime-main.8fa3ef7f.js.map",
-    "js/2.d171c1f4.chunk.js": "/static/react/js/2.d171c1f4.chunk.js",
-    "js/2.d171c1f4.chunk.js.map": "/static/react/js/2.d171c1f4.chunk.js.map",
-    "../../templates/index.html": "/static/react/../../templates/index.html",
-    "css/main.5f361e03.chunk.css.map": "/static/react/css/main.5f361e03.chunk.css.map",
-    "js/2.d171c1f4.chunk.js.LICENSE.txt": "/static/react/js/2.d171c1f4.chunk.js.LICENSE.txt",
-    "precache-manifest.ef6dcd28f1e9e7e985db6e3e26ea3052.js": "/static/react/precache-manifest.ef6dcd28f1e9e7e985db6e3e26ea3052.js",
-    "service-worker.js": "/static/react/service-worker.js"
-  },
-  "entrypoints": [
-    "js/runtime-main.8fa3ef7f.js",
-    "js/2.d171c1f4.chunk.js",
-    "css/main.5f361e03.chunk.css",
-    "js/main.29a4e32b.chunk.js"
-  ]
-}
