--- flask-backend/static/react/precache-manifest.ef6dcd28f1e9e7e985db6e3e26ea3052.js
@@ -1,26 +1,0 @@
-self.__precacheManifest = (self.__precacheManifest || []).concat([
-  {
-    "revision": "3095b7388c2cf913b5715d26cf2e1f0a",
-    "url": "/static/react/../../templates/index.html"
-  },
-  {
-    "revision": "88b41b90c581fb26b8a9",
-    "url": "/static/react/css/main.5f361e03.chunk.css"
-  },
-  {
-    "revision": "638e9c0c00f85b0f791b",
-    "url": "/static/react/js/2.d171c1f4.chunk.js"
-  },
-  {
-    "revision": "e88a3e95b5364d46e95b35ae8c0dc27d",
-    "url": "/static/react/js/2.d171c1f4.chunk.js.LICENSE.txt"
-  },
-  {
-    "revision": "88b41b90c581fb26b8a9",
-    "url": "/static/react/js/main.29a4e32b.chunk.js"
-  },
-  {
-    "revision": "642bffd405cc52de04d1",
-    "url": "/static/react/js/runtime-main.8fa3ef7f.js"
-  }
-]);
